+++ server/src/modules/group.module.js
@@ -2,9 +2,9 @@ import mongoose from "mongoose";
 
 const groupSchema = new mongoose.Schema({
     groupName: { type: String, required: true },
-    admin: { type: Schema.Types.ObjectId, ref: "User", required: true }, 
-    members: [{ type: Schema.Types.ObjectId, ref: "User" }], 
-    messages: [{ type: Schema.Types.ObjectId, ref: "GroupMessage" }], 
+    admin: { type: mongoose.Schema.Types.ObjectId, ref: "User", required: true }, 
+    members: [{ type: mongoose.Schema.Types.ObjectId, ref: "User" }], 
+    messages: [{ type: mongoose.Schema.Types.ObjectId, ref: "GroupMessage" }], 
   },
   { timestamps: true }
 );